Changed the folder structure for better understanding and clean code.
