+++ src/app/page.tsx
@@ -1,17 +1,16 @@
-import Navbar from "@/components/Navbar";
+import Navbar from "@/components/layout/Navbar";
 import ComingSoon from "../components/ComingSoon";
-import Footer from "@/components/Footer";
+import Footer from "@/components/layout/Footer";
 
 const Home = () => {
 	return (
-		<>	
+		<>
 			<Navbar />
-		<div className="flex justify-center bg-primary-monkeyWhite dark:bg-primary-monkeyBlack">
-			<ComingSoon />
-		</div>
-		<Footer/>
+			<div className="flex justify-center bg-primary-monkeyWhite dark:bg-primary-monkeyBlack">
+				<ComingSoon />
+			</div>
+			<Footer />
 		</>
-
 	);
 };
