+++ src/components/basic/ThemeSwitch.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import Icon from "../Icon";
-
 import { useTheme } from "next-themes";
-import Image from "next/image";
 import { useEffect, useState } from "react";
+import Image from "next/image";
+
+import Icon from "../icon";
 
 export default function ThemeSwitch() {
 	const [mounted, setMounted] = useState(false);
@@ -15,7 +15,7 @@ export default function ThemeSwitch() {
 	if (!mounted)
 		return (
 			<Image
-				src="data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24' fill='ff462e'%3E%3Cpath d='M4 3H20C20.5523 3 21 3.44772 21 4V20C21 20.5523 20.5523 21 20 21H4C3.44772 21 3 20.5523 3 20V4C3 3.44772 3.44772 3 4 3ZM5 5V19H19V5H5Z'%3E%3C/path%3E%3C/svg%3E"
+				src="data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24' fill='red'%3E%3Cpath d='M4 3H20C20.5523 3 21 3.44772 21 4V20C21 20.5523 20.5523 21 20 21H4C3.44772 21 3 20.5523 3 20V4C3 3.44772 3.44772 3 4 3ZM5 5V19H19V5H5Z'%3E%3C/path%3E%3C/svg%3E"
 				width={24}
 				height={24}
 				sizes="24x24"
@@ -26,24 +26,10 @@ export default function ThemeSwitch() {
 		);
 
 	if (resolvedTheme === "dark") {
-		return (
-			<Icon
-				title="Toggle Light"
-				name="RiToggleLine"
-				color="#FFF4ED"
-				onClick={() => setTheme("light")}
-			/>
-		);
+		return <Icon name="RiToggleLine" onClick={() => setTheme("light")} />;
 	}
 
 	if (resolvedTheme === "light") {
-		return (
-			<Icon
-				title="Toggle Dark"
-				name="RiToggleFill"
-				color="#101010"
-				onClick={() => setTheme("dark")}
-			/>
-		);
+		return <Icon name="RiToggleFill" onClick={() => setTheme("dark")} />;
 	}
 }
